requirements.txt
beir==2.0.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
elasticsearch==7.9.1
faiss-cpu==1.7.4
greenlet==3.1.1
h5py==3.12.1
llama-cloud==0.1.2
llama-index @ file:///root/llamaindex/llama_index
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core @ file:///root/llamaindex/llama_index/llama-index-core
llama-index-embeddings-huggingface @ file:///root/llamaindex/llama_index/llama-index-integrations/embeddings/llama-index-embeddings-huggingface
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.15
llama-index-multi-modal-llms-openai==0.2.2
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-faiss @ file:///root/llamaindex/llama_index/llama-index-integrations/vector_stores/llama-index-vector-stores-faiss
llama-parse==0.5.10
marshmallow==3.22.0
minijinja==2.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
pypdf==4.3.1
pytrec_eval==0.5
sentence-transformers==3.2.0
SQLAlchemy==2.0.36
striprtf==0.0.26
tenacity==8.5.0
tensorboard==2.18.0
tensorflow==2.18.0
tf_keras==2.18.0
tokenizers==0.20.1
transformers==4.45.2
typing-inspect==0.9.0
